スケーラーがスケールを絶対設定しないように修正

+++ GgafCore/src/jp/ggaf/core/GgafObject.cpp
@@ -3,7 +3,7 @@
 
 using namespace GgafCore;
 
-int GgafObject::_obj_seq = 0;
+uint64_t GgafObject::_obj_seq = 0;
 GgafObject::GgafObject() {
     _obj_class = 0;
     _id = (_obj_seq++);
